vboot: ec: add VbExEcBatteryCutOff callback.
The new API VbExEcBatteryCutOff was added in vboot and the callback has
to be provided in EC module.

The new cut-off function is planned to be supported only on new device
(i.e., those using Depthcharge) and not in U-Boot so the stub call here
simply returns error.

BRANCH=none
BUG=chromium:601705
TEST=emerge-daisy_spring chromeos-u-boot chromeos-bootimage;
CQ-DEPEND=CL:337602

Change-Id: I9cf6842b6702cb9fd731a0e4180c67a9faa1d283
Signed-off-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/338193
Reviewed-by: Duncan Laurie <dlaurie@chromium.org>