patman: Update to use absolute imports
Use absolute imports in patman so that installing patman is easier and it
can run with its modules being in the 'patman' subdirectory.

Rename patman.py since Python does not like the Python module name being
the same as the module directory.

Add the required path to sys.path() so that running patman locally works.

Change-Id: Ia00f61964f3e38a72f4eef24878c0756cda8bf9e
Signed-off-by: Simon Glass <sjg@chromium.org>
BUG=chromium:1048179
TEST=sudo emerge patman
Then run 'patman'
(tests do not pass as they need quite a bit more work for portage)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/u-boot/+/2037907
Reviewed-by: Mike Frysinger <vapier@chromium.org>